some refinement updates
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@731 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/grid/generation/GridGenerator.hpp
@@ -9,6 +9,7 @@
 #define GRIDGENERATOR_HPP
 
 #include "grid/generation/RefinementFunctor.hpp"
+#include "grid/generation/CoarseningFunctor.hpp"
 
 namespace sg
 {
@@ -37,18 +38,32 @@ public:
 	virtual void regular(size_t level) = 0;
 
 	/**
-	 * Refines a regular grid according to the settings of the RefinementFunctor func.
+	 * Refines a grid according to the settings of the RefinementFunctor func.
 	 *
 	 * @param func pointer to refinement functor
 	 */
 	virtual void refine(RefinementFunctor* func) = 0;
 
 	/**
+	 * Coarsens a  grid according to the settings of the CoarseningFunctor func.
+	 *
+	 * @param func pointer to coarsening functor
+	 */
+	//virtual void coarsen(CoarseningFunctor* func) = 0;
+
+	/**
 	 * Returns the number of points on the grid that can be refined in the next iteration
-	 * 
+	 *
 	 * @return the number of points on the grid that can be refined
 	 */
-	virtual int getNumberOfRefinablePoints() = 0;
+	virtual size_t getNumberOfRefinablePoints() = 0;
+
+	/**
+	 * Returns the number of points on the grid that can be removed in the next iteration
+	 *
+	 * @return the number of points on the grid that can be removed
+	 */
+	//virtual size_t getNumberOfRemoveablePoints() = 0;
 };
 
 }
