+++ src/sgpp/grid/generation/StandardGridGenerator.cpp
@@ -35,7 +35,7 @@ void StandardGridGenerator::refine(RefinementFunctor* func)
 	refine.free_refine(this->storage, func);
 }
 
-int StandardGridGenerator::getNumberOfRefinablePoints()
+size_t StandardGridGenerator::getNumberOfRefinablePoints()
 {
 	HashRefinement refine;
 	return refine.getNumberOfRefinablePoints(this->storage);
