+++ src/sgpp/grid/generation/TrapezoidBoundaryGridGenerator.hpp
@@ -35,7 +35,7 @@ public:
 
 	virtual void regular(size_t level);
 	virtual void refine(RefinementFunctor* func);
-	virtual int getNumberOfRefinablePoints();
+	virtual size_t getNumberOfRefinablePoints();
 
 protected:
 	/// Pointer to the grid's storage object
